elastic-beanstalk/requirements.txt
click==8.1.3
colorama==0.4.6
dnspython==2.3.0
eventlet==0.33.3
Flask==2.2.3
Flask-MQTT==1.1.1
greenlet==2.0.2
importlib-metadata==6.1.0
itsdangerous==2.1.2
Jinja2==3.1.2
MarkupSafe==2.1.2
paho-mqtt==1.6.1
six==1.16.0
Werkzeug==2.2.3
zipp==3.15.0
PyMySQL==1.0.3
